Added JWT validation. Added RBAC, based on JWT roles. api-gateway - refactored Route validation.

+++ useradmin/src/main/java/com/performity/useradmin/UserController.java
@@ -1,43 +1,63 @@
 package com.performity.useradmin;
 
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.nio.file.AccessDeniedException;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 
 @RestController
-@RequestMapping("/v1/users")
+@RequestMapping("/useradmin/v1/users")
 public class UserController {
 
     @Autowired
     private UsersService usersService;
 
     @GetMapping
-    public ResponseEntity<List<User>> getAllUsers() {
+    public ResponseEntity<List<User>> getAllUsers(HttpServletRequest request) throws AccessDeniedException {
+        checkAdminPermission(request.getHeader("userRoles"));
         return new ResponseEntity<List<User>>(usersService.allUsers(), HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<User> getUserByEmail(@PathVariable("id") String email) {
+    public ResponseEntity<User> getUserByEmail(HttpServletRequest request, @PathVariable("id") String email) throws AccessDeniedException {
+        // Allow user to read their own data, but only admins get read other users data.
+        if (!email.equals(request.getHeader("userEmail"))) {
+            checkAdminPermission(request.getHeader("userRoles"));
+        }
         return new ResponseEntity<>(usersService.getUserDetails(email), HttpStatus.OK);
     }
 
     @PostMapping
-    public ResponseEntity<User> createUser(@RequestBody Map<String, Object> payload) {
+    public ResponseEntity<User> createUser(HttpServletRequest request, @RequestBody Map<String, Object> payload) throws AccessDeniedException {
+        checkAdminPermission(request.getHeader("userRoles"));
         return new ResponseEntity<User>(usersService.createUser(payload), HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<User> updateTutorial(@PathVariable("id") String email, @RequestBody User user) {
+    public ResponseEntity<User> updateUser(HttpServletRequest request, @PathVariable("id") String email, @RequestBody User user) throws AccessDeniedException {
+        checkAdminPermission(request.getHeader("userRoles"));
         return new ResponseEntity<>(usersService.updateByEmail(email, user), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> deleteUser(@PathVariable("id") String email) {
+    public ResponseEntity<HttpStatus> deleteUser(HttpServletRequest request, @PathVariable("id") String email) throws AccessDeniedException {
+        checkAdminPermission(request.getHeader("userRoles"));
         usersService.deleteByEmail(email);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+
+    private boolean checkAdminPermission(String roles) throws AccessDeniedException {
+        String[] userRoles = roles.split(", ");
+        boolean isAdmin = Arrays.stream(userRoles).anyMatch(x -> "Admin".equals(x));
+        if (!isAdmin) {
+            throw new AccessDeniedException("Access denied");
+        }
+        return true;
     }
 }
